Fixed flickering while music is playing, somehow
Somehow by having the whole vertical trace / page swap block and palette
change code in a single interrupt disable/enable() block fixes it.

At first I thought it had to do with moving the "wait for vertical
trace" to the start of the loop, maybe this syncs up the frames and
palette changes from the get-go?